added explicit casts for (varagr) TIFFSetField() parameters
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@34615 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/imagtiff.cpp
@@ -343,9 +343,9 @@ bool wxTIFFHandler::SaveFile( wxImage *image, wxOutputStream& stream, bool verbo
             image->HasOption(wxIMAGE_OPTION_RESOLUTIONY) )
     {
         TIFFSetField(tif, TIFFTAG_XRESOLUTION,
-                        image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONX));
+                        (float)image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONX));
         TIFFSetField(tif, TIFFTAG_YRESOLUTION,
-                        image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONY));
+                        (float)image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONY));
     }
 
     int spp = image->GetOptionInt(wxIMAGE_OPTION_SAMPLESPERPIXEL);
